requirements.txt
python-socketio==5.11.2
qiskit-finance==0.4.1
# ml_dtypes>=0.3.1
# tensorstore>=0.1.69
requests>=2.31.0
pytz>=2023.3
python-dotenv>=1.0.0

# Dependency Resolution
protobuf==3.20.3  # Required for TensorFlow compatibility

# Data Processing & I/O
pandas==2.1.4
numpy==1.26.3
pyarrow>=14.0.0
python-dateutil>=2.8.0


# AWS Integration
# boto3>=1.34.0
# botocore>=1.34.0
# s3fs==2023.10.0

# GPU Acceleration
# cudnn==8.9.4.25
# cuda-python==12.2.0
# nvidia-cudnn-cu12==8.9.4.25

# System Utilities
tqdm==4.66.1
joblib==1.3.2
backtesting==0.3.3


# Core Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
xgboost==1.7.6
shap==0.45.0

keras
pyopenssl==23.2.0
